Feat: Add validation to /posts/:postId endpoint
1. src/controllers/posts.ts
- body 또는 params에 올바른 입력값을 전달받지 못할경우 400에러를 반환하는 코드 추가

+++ node_level2_express/src/controllers/posts.ts
@@ -11,6 +11,13 @@ import asyncHandler from "../lib/asyncHandler";
 // 게시글 생성
 export const createPost = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
+    const { user, password, title, content } = req.body;
+
+    if (!user || !password || !title || !content) {
+      return res
+        .status(400)
+        .json({ message: "데이터 형식이 올바르지 않습니다." });
+    }
     const newPost: CreatePostRequest = req.body;
     await PostService.createPost(newPost);
     res.send({ message: "게시글을 생성하였습니다." });
@@ -37,8 +44,15 @@ export const getOnePost = asyncHandler(
 // 특정 게시글 수정
 export const updateOnePost = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
-    const updatePostRequest: UpdatePostRequest = req.body;
+    const { password, title, content } = req.body;
     const postId: number = Number(req.params.postId);
+    if (!password || !title || !content || !postId) {
+      return res
+        .status(400)
+        .json({ message: "데이터 형식이 올바르지 않습니다." });
+    }
+
+    const updatePostRequest: UpdatePostRequest = req.body;
     const post = await PostService.updateOnePost(postId, updatePostRequest);
     res.send({ message: "게시글을 수정하였습니다." });
   }
@@ -49,6 +63,11 @@ export const deleteOnePost = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     const { password } = req.body;
     const postId: number = Number(req.params.postId);
+    if (!password || !postId) {
+      return res
+        .status(400)
+        .json({ message: "데이터 형식이 올바르지 않습니다." });
+    }
     await PostService.deleteOnePost(postId, password);
     res.send({ message: "게시글을 삭제하였습니다." });
   }
